将 userViewModel  调整为 appScopeViewModel
(cherry picked from commit 0b6d1fd3137c7a2ba844f87f295ab7ebe398550e)
