+++ chat/src/main/java/cn/wildfire/chat/app/main/MainActivity.java
@@ -646,7 +646,7 @@ public class MainActivity extends WfcBaseActivity {
 
     private void showUser(String uid) {
 
-        UserViewModel userViewModel = ViewModelProviders.of(this).get(UserViewModel.class);
+        UserViewModel userViewModel = WfcUIKit.getAppScopeViewModel(UserViewModel.class);
         UserInfo userInfo = userViewModel.getUserInfo(uid, true);
         if (userInfo == null) {
             return;
@@ -677,7 +677,7 @@ public class MainActivity extends WfcBaseActivity {
     }
 
     private boolean checkDisplayName() {
-        UserViewModel userViewModel = ViewModelProviders.of(this).get(UserViewModel.class);
+        UserViewModel userViewModel = WfcUIKit.getAppScopeViewModel(UserViewModel.class);
         SharedPreferences sp = getSharedPreferences("wfc_config", Context.MODE_PRIVATE);
         UserInfo userInfo = userViewModel.getUserInfo(userViewModel.getUserId(), false);
         if (userInfo != null && TextUtils.equals(userInfo.displayName, userInfo.mobile)) {
