+++ chat/src/main/java/cn/wildfire/chat/app/main/MeFragment.java
@@ -44,6 +44,7 @@ import cn.wildfire.chat.app.login.model.Result;
 import cn.wildfire.chat.app.login.model.Version;
 import cn.wildfire.chat.app.setting.AccountActivity;
 import cn.wildfire.chat.app.setting.SettingActivity;
+import cn.wildfire.chat.kit.WfcUIKit;
 import cn.wildfire.chat.kit.utils.LocaleUtils;
 import cn.wildfire.chat.kit.conversation.file.FileRecordListActivity;
 import cn.wildfire.chat.kit.favorite.FavoriteListActivity;
@@ -159,7 +160,7 @@ public class MeFragment extends Fragment {
     }
 
     private void init() {
-        userViewModel = ViewModelProviders.of(this).get(UserViewModel.class);
+        userViewModel = WfcUIKit.getAppScopeViewModel(UserViewModel.class);
         userViewModel.getUserInfoAsync(userViewModel.getUserId(), true)
                 .observe(getViewLifecycleOwner(), info -> {
                     userInfo = info;
