+++ uikit/src/main/java/cn/wildfire/chat/kit/conversation/SecretConversationInfoFragment.java
@@ -106,7 +106,7 @@ public class SecretConversationInfoFragment extends Fragment implements Conversa
 
     private void init() {
         conversationViewModel = WfcUIKit.getAppScopeViewModel(ConversationViewModel.class);
-        userViewModel = ViewModelProviders.of(this).get(UserViewModel.class);
+        userViewModel = WfcUIKit.getAppScopeViewModel(UserViewModel.class);
 
         stickTopSwitchButton.setChecked(conversationInfo.top > 0);
         silentSwitchButton.setChecked(conversationInfo.isSilent);
