+++ uikit/src/main/java/cn/wildfire/chat/kit/conversation/forward/ForwardActivity.java
@@ -19,6 +19,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import cn.wildfire.chat.kit.R;
+import cn.wildfire.chat.kit.WfcUIKit;
 import cn.wildfire.chat.kit.common.OperateResult;
 import cn.wildfire.chat.kit.conversation.pick.PickOrCreateConversationActivity;
 import cn.wildfire.chat.kit.group.GroupViewModel;
@@ -54,7 +55,7 @@ public class ForwardActivity extends PickOrCreateConversationActivity {
             finish();
         }
         forwardViewModel = ViewModelProviders.of(this).get(ForwardViewModel.class);
-        userViewModel = ViewModelProviders.of(this).get(UserViewModel.class);
+        userViewModel = WfcUIKit.getAppScopeViewModel(UserViewModel.class);
         groupViewModel = ViewModelProviders.of(this).get(GroupViewModel.class);
     }
