+++ uikit/src/main/java/cn/wildfire/chat/kit/conversation/message/viewholder/NormalMessageContentViewHolder.java
@@ -445,7 +445,7 @@ public abstract class NormalMessageContentViewHolder extends MessageContentViewH
     }
 
     private void showGroupMemberAlias(Conversation conversation, Message message, String sender) {
-        UserViewModel userViewModel = ViewModelProviders.of(fragment).get(UserViewModel.class);
+        UserViewModel userViewModel = WfcUIKit.getAppScopeViewModel(UserViewModel.class);
         String hideGroupNickName = userViewModel.getUserSetting(UserSettingScope.GroupHideNickname, conversation.target);
         if ((!TextUtils.isEmpty(hideGroupNickName) && "1".equals(hideGroupNickName)) || message.direction == MessageDirection.Send) {
             nameTextView.setVisibility(View.GONE);
