+++ uikit/src/main/java/cn/wildfire/chat/kit/group/manage/GroupManagerListFragment.java
@@ -17,6 +17,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 
 import cn.wildfire.chat.kit.R;
+import cn.wildfire.chat.kit.WfcUIKit;
 import cn.wildfire.chat.kit.contact.BaseUserListFragment;
 import cn.wildfire.chat.kit.contact.UserListAdapter;
 import cn.wildfire.chat.kit.contact.model.FooterValue;
@@ -72,7 +73,7 @@ public class GroupManagerListFragment extends BaseUserListFragment {
 
     @Override
     public void onUserClick(UIUserInfo userInfo) {
-        UserViewModel userViewModel = ViewModelProviders.of(this).get(UserViewModel.class);
+        UserViewModel userViewModel = WfcUIKit.getAppScopeViewModel(UserViewModel.class);
         GroupMember me = groupViewModel.getGroupMember(groupInfo.target, userViewModel.getUserId());
         if (me == null || me.type != GroupMember.GroupMemberType.Owner) {
             return;
