+++ uikit/src/main/java/cn/wildfire/chat/kit/group/manage/GroupMemberMuteOrAllowListFragment.java
@@ -18,6 +18,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 
 import cn.wildfire.chat.kit.R;
+import cn.wildfire.chat.kit.WfcUIKit;
 import cn.wildfire.chat.kit.contact.BaseUserListFragment;
 import cn.wildfire.chat.kit.contact.UserListAdapter;
 import cn.wildfire.chat.kit.contact.model.HeaderValue;
@@ -78,7 +79,7 @@ public class GroupMemberMuteOrAllowListFragment extends BaseUserListFragment imp
 
     @Override
     public void onUserClick(UIUserInfo userInfo) {
-        UserViewModel userViewModel = ViewModelProviders.of(this).get(UserViewModel.class);
+        UserViewModel userViewModel = WfcUIKit.getAppScopeViewModel(UserViewModel.class);
         GroupMember me = groupViewModel.getGroupMember(groupInfo.target, userViewModel.getUserId());
         if (me == null || me.type != GroupMember.GroupMemberType.Owner) {
             return;
